- Address the JP2 issue of remotes that use 3 bytes to store EFC values > 65535.  Update version to v2.02 Beta 1.5q. (Graham)
git-svn-id: https://svn.code.sf.net/p/controlremote/code/trunk@1109 bfc585dc-5255-483b-a5ca-a9bf5897b33d

+++ km/com/hifiremote/jp1/EFC.java
@@ -84,7 +84,7 @@ public class EFC implements Comparable< EFC >
    *          the hex
    * @return the short
    */
-  public static short parseHex( Hex hex )
+  public static int parseHex( Hex hex )
   {
     return parseHex( hex, 0 );
   }
@@ -98,12 +98,12 @@ public class EFC implements Comparable< EFC >
    *          the index
    * @return the short
    */
-  public static short parseHex( Hex hex, int index )
+  public static int parseHex( Hex hex, int index )
   {
     return parseHex( hex.getData()[ index ] );
   }
 
-  public static short parseHex( short hex )
+  public static int parseHex( short hex )
   {
     short rc = ( short )( hex & 0xFF );
     rc = ( short )( rc << 3 | rc >> 5 );
